Refine value intrinsic tests
