+++ Tests/XMLCoderTests/AttributedIntrinsicTest.swift
@@ -14,7 +14,15 @@ let fooXML = """
 <foo id="123">456</foo>
 """.data(using: .utf8)!
 
-private struct Foo: Codable, DynamicNodeEncoding {
+let fooArrayXML = """
+<?xml version="1.0" encoding="UTF-8"?>
+<container>
+<foo id="123">456</foo>
+<foo id="789">123</foo>
+</container>
+""".data(using: .utf8)!
+
+private struct Foo: Codable, DynamicNodeEncoding, Equatable {
     let id: String
     let value: String
 
@@ -33,7 +41,7 @@ private struct Foo: Codable, DynamicNodeEncoding {
     }
 }
 
-private struct FooEmptyKeyed: Codable, DynamicNodeEncoding {
+private struct FooEmptyKeyed: Codable, DynamicNodeEncoding, Equatable {
     let id: String
     let unkeyedValue: Int
 
@@ -50,6 +58,10 @@ private struct FooEmptyKeyed: Codable, DynamicNodeEncoding {
             return .element
         }
     }
+}
+
+private struct Container<T>: Codable, Equatable where T: Codable & Equatable {
+    let foo: [T]
 }
 
 private let previewXML =
@@ -136,125 +148,29 @@ final class AttributedIntrinsicTest: XCTestCase {
         ), preview)
     }
 
+    func testArray() throws {
+        let decoder = XMLDecoder()
+        decoder.errorContextLength = 10
+
+        let foo1 = try decoder.decode(Container<Foo>.self, from: fooArrayXML)
+        XCTAssertEqual(foo1, Container(foo: [
+            Foo(id: "123", value: "456"),
+            Foo(id: "789", value: "123"),
+        ]))
+
+        let foo2 = try decoder.decode(
+            Container<FooEmptyKeyed>.self,
+            from: fooArrayXML
+        )
+        XCTAssertEqual(foo2, Container(foo: [
+            FooEmptyKeyed(id: "123", unkeyedValue: 456),
+            FooEmptyKeyed(id: "789", unkeyedValue: 123),
+        ]))
+    }
+
     static var allTests = [
         ("testEncode", testEncode),
         ("testDecode", testDecode),
         ("testDecodePreview", testDecodePreview),
-    ]
-}
-
-// MARK: - Enums
-
-let attributedEnumXML = """
-<?xml version="1.0" encoding="UTF-8"?>
-<foo><number type="string">ABC</number><number type="int">123</number></foo>
-""".data(using: .utf8)!
-
-private struct Foo2: Codable {
-    let number: [FooNumber]
-}
-
-private struct FooNumber: Codable, DynamicNodeEncoding {
-    public let type: FooEnum
-
-    public init(type: FooEnum) {
-        self.type = type
-    }
-
-    enum CodingKeys: String, CodingKey {
-        case type
-        case typeValue = ""
-    }
-
-    public static func nodeEncoding(for key: CodingKey) -> XMLEncoder.NodeEncoding {
-        switch key {
-        case FooNumber.CodingKeys.type: return .attribute
-        default: return .element
-        }
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-
-        type = try container.decode(FooEnum.self, forKey: .type)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        switch type {
-        case let .string(value):
-            try container.encode("string", forKey: .type)
-            try container.encode(value, forKey: .typeValue)
-        case let .int(value):
-            try container.encode("int", forKey: .type)
-            try container.encode(value, forKey: .typeValue)
-        }
-    }
-}
-
-private enum FooEnum: Equatable, Codable {
-    private enum CodingKeys: String, CodingKey {
-        case string
-        case int
-    }
-
-    public init(from decoder: Decoder) throws {
-        let values = try decoder.container(keyedBy: CodingKeys.self)
-        if let value = try values.decodeIfPresent(String.self, forKey: .string) {
-            self = .string(value)
-            return
-        } else if let value = try values.decodeIfPresent(Int.self, forKey: .int) {
-            self = .int(value)
-            return
-        } else {
-            throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: decoder.codingPath,
-                                                                    debugDescription: "No coded value for string or int"))
-        }
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        switch self {
-        case let .string(value):
-            try container.encode(value, forKey: .string)
-        case let .int(value):
-            try container.encode(value, forKey: .int)
-        }
-    }
-
-    case string(String)
-    case int(Int)
-}
-
-final class AttributedEnumIntrinsicTest: XCTestCase {
-    func testEncode() throws {
-        let encoder = XMLEncoder()
-        encoder.outputFormatting = []
-
-        let foo1 = Foo2(number: [FooNumber(type: FooEnum.string("ABC")), FooNumber(type: FooEnum.int(123))])
-
-        let header = XMLHeader(version: 1.0, encoding: "UTF-8")
-        let encoded = try encoder.encode(foo1, withRootKey: "foo", header: header)
-        let xmlString = String(data: encoded, encoding: .utf8)
-        XCTAssertNotNil(xmlString)
-        // Test string equivalency
-        let encodedXML = xmlString!.trimmingCharacters(in: .whitespacesAndNewlines)
-        let originalXML = String(data: attributedEnumXML, encoding: .utf8)!.trimmingCharacters(in: .whitespacesAndNewlines)
-        XCTAssertEqual(encodedXML, originalXML)
-    }
-
-    // TODO: Fix decoding
-//    func testDecode() throws {
-//        let decoder = XMLDecoder()
-//        decoder.errorContextLength = 10
-//
-//        let foo = try decoder.decode(Foo2.self, from: attributedEnumXML)
-//        XCTAssertEqual(foo.number[0].type, FooEnum.string("ABC"))
-//        XCTAssertEqual(foo.number[1].type, FooEnum.int(123))
-//    }
-
-    static var allTests = [
-        ("testEncode", testEncode),
-//        ("testDecode", testDecode),
     ]
 }
